feat: prd-classifier-workstation-002 - Return one plain string classification label from execution
