+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -43,6 +43,7 @@ work inputs.
 - When a public workstation success-route field becomes optional or a new authored route family is added, audit downstream generated-contract consumers for required-array assumptions; `outputs` optionality ripples into service validation, replay artifacts, topology projections, and UI topology helpers even if the config mapper change itself is small.
 - Classifier output normalization belongs in `pkg/workers/workstation_executor.go` at the workstation boundary, before transition routing runs; keep plain-string label trimming and invalid-output failure handling there so classifier semantics do not depend on worker stop-token or rejection behavior.
 - Preserve classifier invalid-output evidence at that same workstation boundary by shaping `WorkResult.Error` there; dispatch events, failure details, and projections already inherit failed-workstation error text, so one boundary fix keeps raw returned label evidence aligned across runtime surfaces.
+- When runtime classifier success trims the returned label before route matching, keep authored `classificationRoutes[*].label` canonical too by rejecting leading or trailing whitespace in `pkg/config/config_validator_topology.go`; otherwise authored labels can validate yet become unmatchable at runtime.
 - Classifier route selection spans `pkg/config/config_mapper.go`, `pkg/petri/arc.go`, `pkg/factory/subsystems/subsystem_transitioner.go`, and `pkg/factory/state/validation/type_alignment.go`: map `classificationRoutes` onto labeled accepted arcs, treat those labels as mutually exclusive during accepted routing and same-type output validation, and leave unlabeled output arcs such as resource releases unconditional.
 - When a runtime-only classifier success detail must survive replay or dashboard reconstruction, add it to the canonical `DISPATCH_RESPONSE` payload in `pkg/factory/event_history.go`, decode it in `pkg/replay/event_reducer.go` and `pkg/factory/projections/world_state_generated.go`, and mirror it into `pkg/api/workstation_request_projection.go`; selected-tick projections cannot recover fields that never entered canonical event history.
 - When a public per-input guard type changes, update the same four seams together: source OpenAPI guard schemas, generated Go and TypeScript clients, `pkg/config` boundary normalization and validation, and any split-workstation frontmatter coverage that proves `AGENTS.md` guard loading.
